+++ src/ResponseBuilder.php
@@ -38,6 +38,15 @@ class ResponseBuilder
 	const DEFAULT_API_CODE_OK = ApiCodeBase::OK;
 
 	/**
+	 * Default JSON encoding options
+	 *
+	 * 15 = JSON_HEX_TAG|JSON_HEX_APOS|JSON_HEX_AMP|JSON_HEX_QUOT
+	 *
+	 * This must be as int due to const limits in PHP disallowing expressions.
+	 */
+	const DEFAULT_ENCODING_OPTIONS = 15;
+
+	/**
 	 * Reads and validates "classes" config mapping
 	 *
 	 * @return array|null Classes mapping as specified in configuration or @null if no such config found
@@ -50,7 +59,7 @@ class ResponseBuilder
 
 		if ($classes !== null) {
 			if (!is_array($classes)) {
-				throw new \RuntimeException(sprintf('CONFIG: "classes" mapping must be an array (\'%s\' given)', gettype($classes)));
+				throw new \RuntimeException(sprintf('CONFIG: "classes" mapping must be an array (%s given)', gettype($classes)));
 			}
 
 			$mandatory_keys = ['key',
@@ -59,7 +68,7 @@ class ResponseBuilder
 			foreach ($classes as $class_name => $class_config) {
 				foreach ($mandatory_keys as $key_name) {
 					if (!array_key_exists($key_name, $class_config)) {
-						throw new \RuntimeException("CONFIG: Missing '{$key_name}' for '{$class_name}' mapping");
+						throw new \RuntimeException(sprintf("CONFIG: Missing '%s' for '%s' class mapping", $key_name, $class_name));
 					}
 				}
 			}
@@ -144,20 +153,22 @@ class ResponseBuilder
 	/**
 	 * Returns success
 	 *
-	 * @param mixed|null   $data      payload to be returned as 'data' node, @null if none
-	 * @param integer|null $api_code  API code to be returned with the response or @null for default value
-	 * @param array|null   $lang_args array of arguments passed to Lang if message associated with api_code uses placeholders
-	 * @param integer|null $http_code HTTP return code to be set for this response or @null for default (200)
+	 * @param mixed|null   $data             payload to be returned as 'data' node, @null if none
+	 * @param integer|null $api_code         API code to be returned with the response or @null for default value
+	 * @param array|null   $lang_args        array of arguments passed to Lang if message associated with api_code uses placeholders
+	 * @param integer|null $http_code        HTTP return code to be set for this response or @null for default (200)
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
+	 *
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 */
-	public static function success($data = null, $api_code = null, array $lang_args = null, $http_code = null)
+	public static function success($data = null, $api_code = null, array $lang_args = null, $http_code = null, $encoding_options = null)
 	{
 		if ($api_code === null) {
 			$api_code = static::DEFAULT_API_CODE_OK;
 		}
 
-		return static::buildSuccessResponse($data, $api_code, $lang_args, $http_code);
+		return static::buildSuccessResponse($data, $api_code, $lang_args, $http_code, $encoding_options);
 	}
 
 	/**
@@ -193,17 +204,18 @@ class ResponseBuilder
 	}
 
 	/**
-	 * @param mixed|null   $data      payload to be returned as 'data' node, @null if none
-	 * @param integer|null $api_code  numeric code to be returned as 'code' or null for default value
-	 * @param array|null   $lang_args array of arguments passed to Lang if message associated with api_code uses placeholders
-	 * @param integer|null $http_code HTTP return code to be set for this response
+	 * @param mixed|null   $data             payload to be returned as 'data' node, @null if none
+	 * @param integer|null $api_code         numeric code to be returned as 'code' or null for default value
+	 * @param array|null   $lang_args        array of arguments passed to Lang if message associated with api_code uses placeholders
+	 * @param integer|null $http_code        HTTP return code to be set for this response
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 *
 	 * @throws \InvalidArgumentException Thrown when provided arguments are invalid.
 	 *
 	 */
-	protected static function buildSuccessResponse($data = null, $api_code = null, array $lang_args = null, $http_code = null)
+	protected static function buildSuccessResponse($data = null, $api_code = null, array $lang_args = null, $http_code = null, $encoding_options = null)
 	{
 		if ($http_code === null) {
 			$http_code = static::DEFAULT_HTTP_CODE_OK;
@@ -213,68 +225,71 @@ class ResponseBuilder
 		}
 
 		if (!is_int($api_code)) {
-			throw new \InvalidArgumentException(sprintf("api_code must be integer ('%s' given)", gettype($api_code)));
+			throw new \InvalidArgumentException(sprintf('api_code must be integer (%s given)', gettype($api_code)));
 		}
 		if (!is_int($http_code)) {
-			throw new \InvalidArgumentException(sprintf("http_code must be integer ('%s' given)", gettype($http_code)));
+			throw new \InvalidArgumentException(sprintf('http_code must be integer (%s given)', gettype($http_code)));
 		} elseif (($http_code < 200) || ($http_code > 299)) {
-			throw new \InvalidArgumentException(sprintf('http_code value is invalid. Must be in range 200-299 inclusive, %d given', $http_code));
+			throw new \InvalidArgumentException(sprintf('Invalid http_code (%d). Must be between 200-299 inclusive', $http_code));
 		}
 
-		return static::make(true, $api_code, $api_code, $data, $http_code, $lang_args);
+		return static::make(true, $api_code, $api_code, $data, $http_code, $lang_args, null, $encoding_options);
 	}
 
 	/**
 	 * Builds error Response object. Supports optional arguments passed to Lang::get() if associated error
 	 * message uses placeholders as well as return data payload
 	 *
-	 * @param integer      $api_code  internal API code to be returned
-	 * @param array|null   $lang_args if array, then this passed as arguments to Lang::get() to build final string.
-	 * @param mixed|null   $data      payload array to be returned in 'data' node or response object
-	 * @param integer|null $http_code optional HTTP status code to be used with this response or @null for default
+	 * @param integer      $api_code         internal API code to be returned
+	 * @param array|null   $lang_args        if array, then this passed as arguments to Lang::get() to build final string.
+	 * @param mixed|null   $data             payload array to be returned in 'data' node or response object
+	 * @param integer|null $http_code        optional HTTP status code to be used with this response or @null for default
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 */
-	public static function error($api_code, array $lang_args = null, $data = null, $http_code = null)
+	public static function error($api_code, array $lang_args = null, $data = null, $http_code = null, $encoding_options = null)
 	{
-		return static::buildErrorResponse($data, $api_code, $http_code, $lang_args);
+		return static::buildErrorResponse($data, $api_code, $http_code, $lang_args, $encoding_options);
 	}
 
 	/**
-	 * @param integer    $api_code  numeric code to be returned as 'code'
-	 * @param mixed|null $data      payload to be returned as 'data' node, @null if none
-	 * @param array|null $lang_args |null optional array with arguments passed to Lang::get()
+	 * @param integer      $api_code         numeric code to be returned as 'code'
+	 * @param mixed|null   $data             payload to be returned as 'data' node, @null if none
+	 * @param array|null   $lang_args        optional array with arguments passed to Lang::get()
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 */
-	public static function errorWithData($api_code, $data, array $lang_args = null)
+	public static function errorWithData($api_code, $data, array $lang_args = null, $encoding_options = null)
 	{
-		return static::buildErrorResponse($data, $api_code, null, $lang_args);
+		return static::buildErrorResponse($data, $api_code, null, $lang_args, $encoding_options);
 	}
 
 	/**
-	 * @param integer    $api_code  numeric code to be returned as 'code'
-	 * @param mixed|null $data      payload to be returned as 'data' node, @null if none
-	 * @param integer    $http_code HTTP error code to be returned with this Cannot be @null
-	 * @param array|null $lang_args |null optional array with arguments passed to Lang::get()
+	 * @param integer      $api_code         numeric code to be returned as 'code'
+	 * @param mixed|null   $data             payload to be returned as 'data' node, @null if none
+	 * @param integer      $http_code        HTTP error code to be returned with this Cannot be @null
+	 * @param array|null   $lang_args        optional array with arguments passed to Lang::get()
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 *
 	 * @throws \InvalidArgumentException if http_code is @null
 	 */
-	public static function errorWithDataAndHttpCode($api_code, $data, $http_code, array $lang_args = null)
+	public static function errorWithDataAndHttpCode($api_code, $data, $http_code, array $lang_args = null, $encoding_options = null)
 	{
 		if ($http_code === null) {
 			throw new \InvalidArgumentException('http_code cannot be null. Use errorWithData() instead');
 		}
 
-		return static::buildErrorResponse($data, $api_code, $http_code, $lang_args);
+		return static::buildErrorResponse($data, $api_code, $http_code, $lang_args, $encoding_options);
 	}
 
 	/**
 	 * @param integer    $api_code  numeric code to be returned as 'code'
 	 * @param integer    $http_code HTTP return code to be set for this response or @null for default
-	 * @param array|null $lang_args |null optional array with arguments passed to Lang::get()
+	 * @param array|null $lang_args optional array with arguments passed to Lang::get()
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 *
@@ -290,16 +305,17 @@ class ResponseBuilder
 	}
 
 	/**
-	 * @param integer      $api_code      numeric code to be returned as 'code'
-	 * @param string       $error_message custom message to be returned as part of error response
-	 * @param mixed|null   $data          payload to be returned as 'data' node, @null if none
-	 * @param integer|null $http_code     optional HTTP status code to be used with this response or @null for defaults
+	 * @param integer      $api_code         numeric code to be returned as 'code'
+	 * @param string       $error_message    custom message to be returned as part of error response
+	 * @param mixed|null   $data             payload to be returned as 'data' node, @null if none
+	 * @param integer|null $http_code        optional HTTP status code to be used with this response or @null for defaults
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 */
-	public static function errorWithMessageAndData($api_code, $error_message, $data, $http_code = null)
+	public static function errorWithMessageAndData($api_code, $error_message, $data, $http_code = null, $encoding_options = null)
 	{
-		return static::buildErrorResponse($data, $api_code, $http_code, null, $error_message);
+		return static::buildErrorResponse($data, $api_code, $http_code, null, $error_message, $encoding_options);
 	}
 
 	/**
@@ -318,12 +334,13 @@ class ResponseBuilder
 	 * Builds error Response object. Supports optional arguments passed to Lang::get() if associated error message
 	 * uses placeholders as well as return data payload
 	 *
-	 * @param mixed|null   $data      payload array to be returned in 'data' node or response object
-	 * @param integer      $api_code  internal API code value to be returned
-	 * @param integer|null $http_code optional HTTP status code to be used with this response or @null for default
-	 * @param array|null   $lang_args if array, then this passed as arguments to Lang::get() to build final string.
-	 * @param string|null  $message   custom message to be returned as part of error response
-	 * @param array|null   $headers   optional HTTP headers to be returned in error response
+	 * @param mixed|null   $data             payload array to be returned in 'data' node or response object
+	 * @param integer      $api_code         internal API code value to be returned
+	 * @param integer|null $http_code        optional HTTP status code to be used with this response or @null for default
+	 * @param array|null   $lang_args        if array, then this passed as arguments to Lang::get() to build final string.
+	 * @param string|null  $message          custom message to be returned as part of error response
+	 * @param array|null   $headers          optional HTTP headers to be returned in error response
+	 * @param integer|null $encoding_options see http://php.net/manual/en/function.json-encode.php or @null to use config's value or defaults
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 *
@@ -331,20 +348,21 @@ class ResponseBuilder
 	 *
 	 * @noinspection MoreThanThreeArgumentsInspection
 	 */
-	protected static function buildErrorResponse($data, $api_code, $http_code, $lang_args = null, $message = null, $headers = null)
+	protected static function buildErrorResponse($data, $api_code, $http_code, $lang_args = null, $message = null,
+	                                             $headers = null, $encoding_options = null)
 	{
 		if ($http_code === null) {
 			$http_code = static::DEFAULT_HTTP_CODE_ERROR;
 		}
 
 		if (!is_int($api_code)) {
-			throw new \InvalidArgumentException(sprintf("api_code must be integer ('%s' given)", gettype($api_code)));
+			throw new \InvalidArgumentException(sprintf('api_code must be integer (%s given)', gettype($api_code)));
 		} elseif ($api_code === static::DEFAULT_API_CODE_OK) {
-			throw new \InvalidArgumentException(sprintf('api_code must not be equal to DEFAULT_API_CODE_OK (%d)', static::DEFAULT_API_CODE_OK));
+			throw new \InvalidArgumentException(sprintf('api_code must not be %d (DEFAULT_API_CODE_OK)', static::DEFAULT_API_CODE_OK));
 		} elseif ((!is_array($lang_args)) && ($lang_args !== null)) {
-			throw new \InvalidArgumentException(sprintf("lang_args must be either array or null ('%s' given)", gettype($lang_args)));
+			throw new \InvalidArgumentException(sprintf('lang_args must be either array or null (%s given)', gettype($lang_args)));
 		} elseif (!is_int($http_code)) {
-			throw new \InvalidArgumentException(sprintf("http_code must be integer ('%s' given)", gettype($http_code)));
+			throw new \InvalidArgumentException(sprintf('http_code must be integer (%s given)', gettype($http_code)));
 		} elseif ($http_code < 400) {
 			throw new \InvalidArgumentException('http_code cannot be lower than 400');
 		}
@@ -356,7 +374,7 @@ class ResponseBuilder
 			$headers = [];
 		}
 
-		return static::make(false, $api_code, $message, $data, $http_code, $lang_args, $headers);
+		return static::make(false, $api_code, $message, $data, $http_code, $lang_args, $headers, $encoding_options);
 	}
 
 
@@ -366,8 +384,9 @@ class ResponseBuilder
 	 * @param string|integer $message_or_api_code message string or API code
 	 * @param mixed|null     $data                optional additional data to be included in response object
 	 * @param integer|null   $http_code           return HTTP code for build Response object
-	 * @param array          $lang_args           |null optional array with arguments passed to Lang::get()
-	 * @param array          $headers             |null optional HTTP headers to be returned in the response
+	 * @param array|null     $lang_args           optional array with arguments passed to Lang::get()
+	 * @param array|null     $headers             optional HTTP headers to be returned in the response
+	 * @param integer|null   $encoding_options    see http://php.net/manual/en/function.json-encode.php
 	 *
 	 * @return \Symfony\Component\HttpFoundation\Response
 	 *
@@ -376,7 +395,9 @@ class ResponseBuilder
 	 * @noinspection MoreThanThreeArgumentsInspection
 	 */
 	protected static function make($success, $api_code, $message_or_api_code, $data = null,
-	                               $http_code = null, array $lang_args = null, array $headers = null)
+	                               $http_code = null, array $lang_args = null, array $headers = null,
+	                               $encoding_options = null
+	)
 	{
 		if ($lang_args === null) {
 			$lang_args = ['api_code' => $message_or_api_code];
@@ -390,12 +411,19 @@ class ResponseBuilder
 				: static::DEFAULT_HTTP_CODE_ERROR;
 		}
 
+		if ($encoding_options === null) {
+			$encoding_options = Config::get('response_builder.encoding_options', static::DEFAULT_ENCODING_OPTIONS);
+		}
+		if (!is_int($encoding_options)) {
+			throw new \InvalidArgumentException(sprintf('encoding_options must be integer (%s given)', gettype($encoding_options)));
+		}
+
 		// are we given message text already?
 		if (!is_string($message_or_api_code)) {
 			// no, so it must be an int value
 			if (!is_int($message_or_api_code)) {
 				throw new \InvalidArgumentException(
-					sprintf('Message must be either string or resolvable integer api_code (\'%s\' given)', gettype($message_or_api_code))
+					sprintf('Message must be either string or resolvable integer api_code (%s given)', gettype($message_or_api_code))
 				);
 			}
 
@@ -412,17 +440,17 @@ class ResponseBuilder
 		} else {
 			if (!is_int($api_code)) {
 				throw new \InvalidArgumentException(
-					sprintf("api_code must be integer ('%s' given)", gettype($api_code))
+					sprintf('api_code must be integer (%s given)', gettype($api_code))
 				);
 			}
 
 			if (!ApiCodeBase::isCodeValid($api_code)) {
-				$msg = sprintf("API code value ({$api_code}) is out of allowed range %d-%d",
-					ApiCodeBase::getMinCode(), ApiCodeBase::getMaxCode());
+				$msg = sprintf('API code value (%d) is out of allowed range %d-%d',
+					$api_code, ApiCodeBase::getMinCode(), ApiCodeBase::getMaxCode());
 				throw new \InvalidArgumentException($msg);
 			}
 		}
 
-		return Response::json(static::buildResponse($success, $api_code, $message_or_api_code, $data), $http_code, $headers);
+		return Response::json(static::buildResponse($success, $api_code, $message_or_api_code, $data), $http_code, $headers, $encoding_options);
 	}
 }