+++ tests/Base/TestCaseBase.php
@@ -125,4 +125,67 @@ abstract class TestCaseBase extends \Orchestra\Testbench\TestCase
 		return $prefix . md5(uniqid(mt_rand(), true));
 	}
 
+
+	/**
+	 * UTF8 aware version of ord()
+	 *
+	 * @param string $string UTF8 string to work on
+	 * @param int    $offset start offset. Note, offset will be updated to properly skip multibyte chars!
+	 *
+	 * $text = "abcàêß€abc";
+	 * $offset = 0;
+	 * while ($offset >= 0) {
+	 *    printf("%d: %d\n", $offset, ord8($text, $offset));
+	 * }
+	 *
+	 * @return int code of the character
+	 */
+	protected function ord8($string, &$offset)
+	{
+		$code = ord(substr($string, $offset, 1));
+		if ($code >= 128) {             //otherwise 0xxxxxxx
+			if ($code < 224) {          //110xxxxx
+				$bytesnumber = 2;
+			} elseif ($code < 240) {    //1110xxxx
+				$bytesnumber = 3;
+			} elseif ($code < 248) {    //11110xxx
+				$bytesnumber = 4;
+			}
+
+			$codetemp = $code - 192 - ($bytesnumber > 2 ? 32 : 0) - ($bytesnumber > 3 ? 16 : 0);
+			for ($i = 2; $i <= $bytesnumber; $i++) {
+				$offset++;
+				$code2 = ord(substr($string, $offset, 1)) - 128;        //10xxxxxx
+				$codetemp = $codetemp * 64 + $code2;
+			}
+			$code = $codetemp;
+		}
+		$offset += 1;
+		if ($offset >= strlen($string)) {
+			$offset = -1;
+		}
+
+		return $code;
+	}
+
+	/**
+	 * UTF8 escape of given string
+	 *
+	 * @param string $string UTF8 string to escape
+	 *
+	 * @return string
+	 */
+	protected function escape8($string)
+	{
+		$escaped = '';
+
+		// escape UTF8 for further comparision
+		$offset = 0;
+		while ($offset >= 0) {
+			$escaped .= sprintf('\u%04x', $this->ord8($string, $offset));
+		}
+
+		return $escaped;
+	}
+
 }